Adds Requirements Model and Views to traq
Requirements are meant to be detailed deliverables. They can be approved by a
client (a user in group 'arcclient') to show that the client understands and
approves the implementation. Requirements are associated with a project and
can optionally be attached to a milestone. Commit adds new app, models,
migrations, List, Detail, Create, Update, and Approve Views and associated templates.